+++ .eslintrc.js
@@ -21,6 +21,13 @@ module.exports = {
     '@typescript-eslint/interface-name-prefix': 'off',
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
-    'import/order': 'warn',
+    'import/order': [
+      'warn',
+      {
+        alphabetize: {
+          order: 'asc',
+        },
+      },
+    ],
   },
 }