+++ src/errors/index.ts
@@ -1,2 +1,4 @@
+export * from './DataSourceImportationError'
 export * from './DataSourceNotProvidedError'
 export * from './SeederImportationError'
+export * from './SeederExecutionError'