+++ test/fixtures/Pet.factory.ts
@@ -1,17 +1,17 @@
 import { faker } from '@faker-js/faker'
-import { LazyInstanceAttribute } from '../../src'
-import { Factory } from '../../src/factory'
-import { Subfactory } from '../../src/subfactory'
-import type { FactorizedAttrs } from '../../src/types'
+import { FactorizedAttrs, Factory, LazyInstanceAttribute, SingleSubfactory } from '@jorgebodega/typeorm-factory'
 import { Pet } from './Pet.entity'
 import { UserFactory } from './User.factory'
+import { dataSource } from './dataSource'
 
 export class PetFactory extends Factory<Pet> {
   protected entity = Pet
+  protected dataSource = dataSource
+
   protected attrs(): FactorizedAttrs<Pet> {
     return {
       name: faker.animal.insect(),
-      owner: new LazyInstanceAttribute((instance) => new Subfactory(UserFactory, { pets: [instance] })),
+      owner: new LazyInstanceAttribute((instance) => new SingleSubfactory(UserFactory, { pets: [instance] })),
     }
   }
 }