+++ test/helpers/useSeeders.test.ts
@@ -1,9 +1,9 @@
 import { useDataSource, useSeeders } from '../../src'
-import { dataSource } from '../fixtures/dataSource'
 import { Pet } from '../fixtures/Pet.entity'
 import PetSeeder from '../fixtures/Pet.seeder'
 import { User } from '../fixtures/User.entity'
 import UserSeeder from '../fixtures/User.seeder'
+import { dataSource } from '../fixtures/dataSource'
 
 describe(useSeeders, () => {
   beforeAll(async () => {
@@ -11,11 +11,7 @@ describe(useSeeders, () => {
   })
 
   beforeEach(async () => {
-    await dataSource.synchronize()
-  })
-
-  afterEach(async () => {
-    await dataSource.dropDatabase()
+    await dataSource.synchronize(true)
   })
 
   afterAll(async () => {